修复投注记录请求当天参数错误、排版错乱、金额保留两位等问题

+++ lib/ui/bean/bet_detail_item_child_entity.dart
@@ -33,7 +33,7 @@ class Record {
   String? tableId;
   String? betamount;
   String? validamount;
-  int? winlose;
+  double? winlose;
   String? betTime;
   String? gameType;
   String? gameName;
@@ -60,7 +60,7 @@ class Record {
 
 @JsonSerializable()
 class Total {
-  int? betAmountTotal;
+  double? betAmountTotal;
   double? validAmountTotal;
   double? winloseTotal;
   int? betCountTotal;
